+++ src/aux-common/Files/FilesChannel.ts
@@ -378,6 +378,7 @@ export type LocalEvents =
     | ShowToastEvent
     | TweenToEvent
     | OpenQRCodeScannerEvent
+    | ShowQRCodeEvent
     | LoadSimulationEvent
     | UnloadSimulationEvent
     | SuperShoutEvent;
@@ -411,12 +412,29 @@ export interface TweenToEvent extends LocalEvent {
  * An event that is used to show or hide the QR Code Scanner.
  */
 export interface OpenQRCodeScannerEvent extends LocalEvent {
-    name: 'show_qr_code';
+    name: 'show_qr_code_scanner';
 
     /**
      * Whether the QR Code scanner should be visible.
      */
     open: boolean;
+}
+
+/**
+ * An event that is used to show or hide a QR Code on screen.
+ */
+export interface ShowQRCodeEvent extends LocalEvent {
+    name: 'show_qr_code';
+
+    /**
+     * Whether the QR Code should be visible.
+     */
+    open: boolean;
+
+    /**
+     * The code to display.
+     */
+    code: string;
 }
 
 /**
@@ -602,8 +620,22 @@ export function tweenTo(fileId: string, zoomValue: number = -1): TweenToEvent {
 export function openQRCodeScanner(open: boolean): OpenQRCodeScannerEvent {
     return {
         type: 'local',
+        name: 'show_qr_code_scanner',
+        open: open,
+    };
+}
+
+/**
+ * Creates a new ShowQRCodeEvent.
+ * @param open Whether the QR Code should be visible.
+ * @param code The code that should be shown.
+ */
+export function showQRCode(open: boolean, code?: string): ShowQRCodeEvent {
+    return {
+        type: 'local',
         name: 'show_qr_code',
         open: open,
+        code: code,
     };
 }
 
@@ -644,3 +676,5 @@ export function superShout(eventName: string, arg?: any): SuperShoutEvent {
         argument: arg,
     };
 }
+
+// export