+++ src/aux-server/aux-web/aux-player/interaction/ClickOperation/PlayerFileClickOperation.ts
@@ -8,11 +8,15 @@ import {
     getFilePosition,
     objectsAtContextGridPosition,
     getFileIndex,
+    duplicateFile,
+    File,
+    getFileDragMode,
 } from '@casual-simulation/aux-common';
 import { BaseFileDragOperation } from '../../../shared/interaction/DragOperation/BaseFileDragOperation';
 import { PlayerFileDragOperation } from '../DragOperation/PlayerFileDragOperation';
 import { dropWhile } from 'lodash';
 import { PlayerSimulation3D } from '../../scene/PlayerSimulation3D';
+import { PlayerNewFileDragOperation } from '../DragOperation/PlayerNewFileDragOperation';
 
 export class PlayerFileClickOperation extends BaseFileClickOperation {
     // This overrides the base class BaseInteractionManager
@@ -20,21 +24,34 @@ export class PlayerFileClickOperation extends BaseFileClickOperation {
     // This overrides the base class IGameView
     protected _simulation3D: PlayerSimulation3D;
 
+    protected faceClicked: { face: string };
+
     constructor(
         simulation: PlayerSimulation3D,
         interaction: PlayerInteractionManager,
-        file: AuxFile3D
+        file: AuxFile3D,
+        faceValue: string
     ) {
         super(simulation, interaction, file.file, file);
+        this.faceClicked = { face: faceValue };
     }
 
     protected _performClick(calc: FileCalculationContext): void {
-        this.simulation.helper.action('onClick', [this._file]);
+        this.simulation.helper.action(
+            'onClick',
+            [this._file],
+            this.faceClicked
+        );
     }
 
     protected _createDragOperation(
         calc: FileCalculationContext
     ): BaseFileDragOperation {
+        const mode = getFileDragMode(calc, this._file);
+        if (mode === 'clone') {
+            return this._createCloneDragOperation();
+        }
+
         const file3D: AuxFile3D = <AuxFile3D>this._file3D;
         const context = file3D.context;
         const position = getFilePosition(calc, file3D.file, context);
@@ -60,5 +77,15 @@ export class PlayerFileClickOperation extends BaseFileClickOperation {
         }
 
         return null;
+    }
+
+    protected _createCloneDragOperation(): BaseFileDragOperation {
+        let duplicatedFile = duplicateFile(<File>this._file);
+        return new PlayerNewFileDragOperation(
+            this._simulation3D,
+            this._interaction,
+            duplicatedFile,
+            (<AuxFile3D>this._file3D).context
+        );
     }
 }