+++ src/aux-server/aux-web/aux-player/interaction/ClickOperation/PlayerInventoryFileClickOperation.ts
@@ -4,6 +4,7 @@ import {
     Object,
     duplicateFile,
     FileCalculationContext,
+    getFileDragMode,
 } from '@casual-simulation/aux-common';
 import { BaseFileDragOperation } from '../../../shared/interaction/DragOperation/BaseFileDragOperation';
 import { BaseFileClickOperation } from '../../../shared/interaction/ClickOperation/BaseFileClickOperation';
@@ -12,6 +13,7 @@ import { PlayerInteractionManager } from '../PlayerInteractionManager';
 import { PlayerFileDragOperation } from '../DragOperation/PlayerFileDragOperation';
 import { InventoryItem } from 'aux-web/aux-player/InventoryContext';
 import { PlayerSimulation3D } from '../../scene/PlayerSimulation3D';
+import { PlayerNewFileDragOperation } from '../DragOperation/PlayerNewFileDragOperation';
 
 /**
  * New File Click Operation handles clicking of files that are in the file queue.
@@ -41,11 +43,28 @@ export class PlayerInventoryFileClickOperation extends BaseFileClickOperation {
         this._item.simulation.simulation.helper.action('onClick', [this._file]);
     }
 
-    protected _createDragOperation(): BaseFileDragOperation {
+    protected _createDragOperation(
+        calc: FileCalculationContext
+    ): BaseFileDragOperation {
+        const mode = getFileDragMode(calc, this._file);
+        if (mode === 'clone') {
+            return this._createCloneDragOperation();
+        }
+
         return new PlayerFileDragOperation(
             this._simulation3D,
             this._interaction,
             [this._file],
+            this._context
+        );
+    }
+
+    protected _createCloneDragOperation(): BaseFileDragOperation {
+        let duplicatedFile = duplicateFile(<File>this._file);
+        return new PlayerNewFileDragOperation(
+            this._simulation3D,
+            this._interaction,
+            duplicatedFile,
             this._context
         );
     }