+++ src/aux-server/aux-web/aux-projector/App/App.ts
@@ -124,6 +124,11 @@ export default class App extends Vue {
     extraItems: SidebarItem[] = [];
 
     /**
+     * The QR Code to display.
+     */
+    qrCode: string = '';
+
+    /**
      * Gets whether we're in developer mode.
      */
     get dev() {
@@ -190,12 +195,18 @@ export default class App extends Vue {
         return location.href;
     }
 
+    getQRCode() {
+        return this.qrCode || this.url();
+    }
+
     currentUserMode() {
         return this.userMode ? 'Files' : 'Worksurfaces';
     }
 
     forcedOffline() {
-        return appManager.simulationManager.primary.socketManager.forcedOffline;
+        return appManager.simulationManager.primary
+            ? appManager.simulationManager.primary.socketManager.forcedOffline
+            : false;
     }
 
     toggleOpen() {
@@ -275,6 +286,14 @@ export default class App extends Vue {
                                 message: e.message,
                                 visible: true,
                             };
+                        } else if (e.name === 'show_qr_code') {
+                            if (e.open) {
+                                this.qrCode = e.code;
+                                this.showQRCode = true;
+                            } else {
+                                this.qrCode = null;
+                                this.showQRCode = false;
+                            }
                         }
                     })
                 );